Addeds different evaluation modes and a way to specify a input text file containing sizes instead of randomly generating them

+++ src/matrix.h
@@ -6,7 +6,7 @@ void setupMatrices(double **A, double **copyA, int *sizes, int n);
 void initializeMatrices (double **A, double **copyA, int *sizes, int n);
 void getAllOrders(int **allOrder, int n);
 void convertOrders(int **allOrder, int n);
-void computeChainCosts(int **allOrder, int *orderCost, int *normSizes, int n, int fac, char cf);
+void computeChainCosts(int **allOrder, double *orderCost, int *normSizes, int n, int fac, char cf);
 void resetInterMatrices(double **interRes, int n);
 void setupInterMatrices(double **interRes, int *order, int *sizes, int n);
